class DynamicAdapt
